src/requirements.txt
hvac
boto3
structlog
PyYAML
